User Profile Update
- redesigned the user page to better match the design of the site
- made the userProfile and the publicUserProfile consistent design-wise
- updated the fields in the user profile to match our specification (added a toggle for profile privacy, removed unused fields)
- added a stats section to the user page
- added an option to preview the public user profile
