+++ app/components/userMenu/achievements.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import React from "react";
+
+const Achievements: React.FC = () => {
+  return <p>This component is currently under construction.</p>;
+};
+
+export default Achievements;
