+++ app/components/userMenu/gameHistory.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import React from "react";
+
+const GameHistory: React.FC = () => {
+  return <p>This component is currently under construction.</p>;
+};
+
+export default GameHistory;
