+++ app/types/user.ts
@@ -1,14 +1,23 @@
 export interface User {
+  userid: number;
   username: string;
   email: string;
   token: string;
-  online: boolean;
+  statsPublic: boolean;
 }
+
+export interface UserAttributes {
+  xp: number;
+  gamesPlayed: number;
+  wins: number;
+  points: number;
+}
+
 export interface PublicUser {
   username: string;
 }
 
-export interface UserAttributes {
+export interface PublicUserAttributes {
   xp: number;
   gamesPlayed: number;
   wins: number;
